requirements.txt
streamlit==1.28.1
anthropic==0.3.11
chromadb==0.4.15
PyPDF2==3.0.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3